+++ lets-fiware.sh
@@ -762,16 +762,18 @@ setup_firewall() {
 install_docker_ubuntu() {
   logging_info "${FUNCNAME[0]}"
 
-  ${SUDO} cp -p /etc/apt/sources.list{,.bak}
   ${APT_GET} update
   ${APT_GET} install -y \
       apt-transport-https \
       ca-certificates \
       curl \
-      gnupg-agent \
-      software-properties-common
-  curl -fsSL https://download.docker.com/linux/ubuntu/gpg | ${APT_KEY} add -
-  ${ADD_APT_REPOSITORY} "deb [arch=amd64] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable"
+      gnupg \
+      lsb-release
+  curl -fsSL https://download.docker.com/linux/ubuntu/gpg | ${SUDO} gpg --dearmor --yes -o /usr/share/keyrings/docker-archive-keyring.gpg
+  echo \
+  "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/docker-archive-keyring.gpg] https://download.docker.com/linux/ubuntu \
+  $(lsb_release -cs) stable" | sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
+  ${APT_GET} update
   ${APT_GET} install -y docker-ce docker-ce-cli containerd.io
   ${SYSTEMCTL} start docker
   ${SYSTEMCTL} enable docker
@@ -800,8 +802,8 @@ check_docker() {
 
   if ! type "${DOCKER_CMD}" >/dev/null 2>&1; then
     case "${DISTRO}" in
-       "Ubuntu" ) install_docker_ubuntu ;;
-       "CentOS" ) install_docker_centos ;;
+      "Ubuntu" ) install_docker_ubuntu ;;
+      "CentOS" ) install_docker_centos ;;
     esac
   fi
 
@@ -825,25 +827,36 @@ check_docker() {
 }
 
 #
-# Check docker-compose
+# Install Docker compose V2 for Ubuntu
+#
+install_docker_compose_ubuntu() {
+  logging_info "${FUNCNAME[0]}"
+
+  ${APT_GET} install -y docker-compose-plugin
+}
+
+#
+# Install Docker compose V2 for CentOS
+#
+install_docker_compose_centos() {
+  logging_info "${FUNCNAME[0]}"
+
+  ${YUM} install -y docker-compose-plugin
+}
+
+#
+# Check docker compose v2
 #
 check_docker_compose() {
   logging_info "${FUNCNAME[0]}"
 
-  if [ -e "${DOCKER_COMPOSE_CMD}" ]; then
-    local ver
-    ver=$(${DOCKER_COMPOSE} --version)
-    logging_info "${FUNCNAME[0]} ${ver}"
-
-    ver=$(${DOCKER_COMPOSE} version --short | awk -F. '{printf "%2d%02d%02d", $1,$2,$3}')
-    if [ "${ver}" -ge 11700 ]; then
-      return
-    fi
+  found=$(sudo docker info --format '{{json . }}' | jq -r '.ClientInfo.Plugins | .[].Name' | grep -ic compose)
+  if [ "${found}" -eq 0 ]; then
+    case "${DISTRO}" in
+      "Ubuntu" ) install_docker_compose_ubuntu ;;
+      "CentOS" ) install_docker_compose_centos ;;
+    esac
   fi
-
-  curl -sOL https://github.com/docker/compose/releases/download/1.29.2/docker-compose-Linux-x86_64
-  ${SUDO} mv docker-compose-Linux-x86_64 "${DOCKER_COMPOSE_CMD}"
-  ${SUDO} chmod a+x "${DOCKER_COMPOSE_CMD}"
 }
 
 #
@@ -3516,8 +3529,7 @@ init_cmd() {
   FIREWALL_CMD="${SUDO} ${MOCK_PATH}firewall-cmd"
   UNAME="${FIBB_TEST_UNAME_CMD:-uname}"
   DOCKER_CMD="${FIBB_TEST_DOCKER_CMD:-docker}"
-  DOCKER_COMPOSE_CMD="/usr/local/bin/docker-compose"
-  DOCKER_COMPOSE="${SUDO} ${DOCKER_COMPOSE_CMD}"
+  DOCKER_COMPOSE="${SUDO} /usr/bin/docker compose"
   HOST_CMD="${FIBB_TEST_HOST_CMD:-host}"
   WAIT_TIME=${FIBB_WAIT_TIME:-300}
   SKIP_INSTALL_WIDGET="${FIBB_TEST_SKIP_INSTALL_WIDGET:-false}"